Fixed an issue where the library used AES-256 instead of AES-128 as intended.
The numbers of the MCRYPT_RIJNDAEL_* constants refer to the block sizes of the Rijndael algorithm, whereas the numbers of the AES variants refer to different key sizes. AES is Rijndael with a block size of 128 bits and a key length of 128 bits (AES-128), 192 bits (AES-192) or 256 bits (AES-256). The library erroneously used MCRYPT_RIJNDAEL_128 together with a 32-byte key instead of the intended 16 byte key.